+++ Example/src/it/polito/verifoo/rest/jaxb/Host.java
@@ -2,7 +2,7 @@
 // Questo file è stato generato dall'architettura JavaTM per XML Binding (JAXB) Reference Implementation, v2.2.8-b130911.1802 
 // Vedere <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Qualsiasi modifica a questo file andrà persa durante la ricompilazione dello schema di origine. 
-// Generato il: 2017.11.15 alle 10:54:36 PM CET 
+// Generato il: 2017.11.16 alle 12:05:13 PM CET 
 //
 
 
@@ -31,7 +31,7 @@ import javax.xml.bind.annotation.XmlType;
  *         &lt;element name="NodeRef" type="{}NodeRefType" maxOccurs="unbounded" minOccurs="0"/>
  *       &lt;/sequence>
  *       &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
- *       &lt;attribute name="diskStorage" use="required" type="{http://www.w3.org/2001/XMLSchema}int" />
+ *       &lt;attribute name="diskStorage" use="required" type="{}positiveFloat" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -51,7 +51,7 @@ public class Host {
     @XmlAttribute(name = "name", required = true)
     protected String name;
     @XmlAttribute(name = "diskStorage", required = true)
-    protected int diskStorage;
+    protected float diskStorage;
 
     /**
      * Gets the value of the nodeRef property.
@@ -110,7 +110,7 @@ public class Host {
      * Recupera il valore della proprietà diskStorage.
      * 
      */
-    public int getDiskStorage() {
+    public float getDiskStorage() {
         return diskStorage;
     }
 
@@ -118,7 +118,7 @@ public class Host {
      * Imposta il valore della proprietà diskStorage.
      * 
      */
-    public void setDiskStorage(int value) {
+    public void setDiskStorage(float value) {
         this.diskStorage = value;
     }
